feat: build columns after pivoting data
