+++ packages/frontend/src/hooks/useColumnTotals.ts
@@ -3,10 +3,11 @@ import {
     DimensionType,
     Field,
     FieldId,
-    getResultColumnTotals,
+    getItemId,
     isAdditionalMetric,
     isField,
     MetricType,
+    ResultRow,
     TableCalculation,
 } from '@lightdash/common';
 import { useMemo } from 'react';
@@ -16,7 +17,7 @@ type Args = {
     itemsMap: Record<FieldId, Field | TableCalculation>;
 };
 
-const isSummable = (item: Field | TableCalculation) => {
+export const isSummable = (item: Field | TableCalculation) => {
     if (isField(item) || isAdditionalMetric(item)) {
         const numericTypes: string[] = [
             DimensionType.NUMBER,
@@ -30,12 +31,46 @@ const isSummable = (item: Field | TableCalculation) => {
     }
     return true;
 };
+
+export const getResultColumnTotals = (
+    rows: ResultRow[],
+    keys: Array<string>,
+): Record<FieldId, number | undefined> => {
+    if (keys.length > 0) {
+        return rows.reduce<Record<FieldId, number | undefined>>(
+            (acc, row: ResultRow) => {
+                keys.forEach((key) => {
+                    if (row[key]) {
+                        acc[key] = (acc[key] || 0) + Number(row[key].value.raw);
+                    }
+                });
+                return acc;
+            },
+            {},
+        );
+    }
+    return {};
+};
+
+export const getResultColumnTotalsFromItemsMap = (
+    rows: ResultRow[],
+    itemsMap: Record<FieldId, Field | TableCalculation>,
+): Record<FieldId, number | undefined> => {
+    return getResultColumnTotals(
+        rows,
+        Object.values(itemsMap).reduce<string[]>(
+            (acc, item) => (isSummable(item) ? [...acc, getItemId(item)] : acc),
+            [],
+        ),
+    );
+};
+
 const useColumnTotals = ({ resultsData, itemsMap }: Args) => {
     return useMemo<Record<FieldId, number | undefined>>(() => {
         if (resultsData) {
-            return getResultColumnTotals(
+            return getResultColumnTotalsFromItemsMap(
                 resultsData.rows,
-                Object.values(itemsMap).filter((field) => isSummable(field)),
+                itemsMap,
             );
         }
         return {};
